feat(nodes-panel): merge dicom nodes dialog into nodes panel
