+++ invesalius/gui/frame.py
@@ -38,7 +38,6 @@ import invesalius.gui.import_panel as imp
 import invesalius.gui.log as log
 import invesalius.gui.preferences as preferences
 import invesalius.gui.dicom_server_panel as dicom_server
-import invesalius.gui.dicom_nodes_panel as dicom_nodes
 
 import invesalius.gui.import_network_panel as imp_net
 import invesalius.project as prj
@@ -738,8 +737,6 @@ class Frame(wx.Frame):
             self.ShowRetrieveDicomPanel()
         elif id == const.ID_DICOM_SERVER:
             self.ShowDicomServer()
-        elif id == const.ID_DICOM_NODE:
-            self.ShowDicomNodes()
         elif id in (const.ID_FLIP_X, const.ID_FLIP_Y, const.ID_FLIP_Z):
             axis = {const.ID_FLIP_X: 2, const.ID_FLIP_Y: 1, const.ID_FLIP_Z: 0}[id]
             self.FlipVolume(axis)
@@ -953,17 +950,6 @@ class Frame(wx.Frame):
             session = ses.Session()
             session.SetConfig('server_aetitle', dicom_server_dialog.ae_title)
             session.SetConfig('server_port', dicom_server_dialog.port)
-
-    def ShowDicomNodes(self):
-        """ Show dicom nodes dialog. """
-
-        dicom_nodes_dialog = dicom_nodes.DicomNodes(self)
-        if dicom_nodes_dialog.ShowModal() == wx.ID_OK:
-
-            dicom_nodes_dialog.Destroy()
-
-            session = ses.Session()
-            session.SetConfig('nodes', dicom_nodes_dialog.nodes)
 
     def ShowPreferences(self, page=0):
         preferences_dialog = preferences.Preferences(self, page)
@@ -1714,8 +1700,6 @@ class MenuBar(wx.MenuBar):
         # NETWORK
         network_settings_menu = wx.Menu()
         network_settings_menu.Append(const.ID_DICOM_SERVER, _("Dicom Server"))
-        network_settings_menu.Append(const.ID_DICOM_NODE, _("Dicom Nodes"))
-        #network_settings_menu.Append(const.ID_DICOM_QUERY, _("Query"))
 
         # OPTIONS
         options_menu = wx.Menu()
